EST-908 (and related): email invoices
This is a rebase (squash) of 33 commits.

EST-881: improves the title for AgreementRole, eg "Tenant of OXF-POISON-123 (Poison Perfumeries)"

EST-887: alters the Binder implementation for prelim doc and invoice note so attach the created document to the invoice's buyer and seller as well as to the invoice itself.

EST-908: removes unnecessary injected service from DocumentTypeMenu

EST-908: updates to latest snapshot of document module (to pick up preview, create/attach, delete actions).

EST-908: extends fixture data for OxfPoison003Gb and PoisonGb so that an invoice address is set up for the lease.

EST-889: extends Invoice#remove so that (if not disallowed due to the state of the invoice itself) then any paperclips for this invoice will be deleted  too.

In other words, a cascade delete of paperclips for temporary invoices.

EST-888: vetoes Document#delete if it has already been sent as a communication.

Also introduced CommunicationRepository (to query for any existing comms), and factored out existing functionality to create comms (from Document_email and Document_postal mixins) and also moved into CommunicationRepository.

EST-908: adds EmailServiceForEstatio, no longer use debugmail.io

EST-908: fixes cover note templates, and improved override of EmailService to provide better error reporting and minimizing the timeout.

EST-908 adds a guard around {invoice.invoiceNumber} in the InvoiceEmailCoverNote html templates (in case used when the invoice is not in state of sent)

ESt-908: no longer allows Invoice Notes to be sent for invoices that do not have an invoiceNumber assigned (that is, which have not been collected)

EST-908: splitting out preview vs create/attach doc

EST-908: references incode-module-unittestsupport rather than incode-module-base for unit test utilities.

EST-908: fixes compile issues in fixture data for DocumentTemplates

EST-887: adds integration test for preliminary letter, confirming attached to invoice, buyer and seller.

Also checks that is rendered and can create a comm

EST-887: also checks that the covernote is created correctly

EST-887: also checks that the covernote is created correctly

EST-888, EST-889 & EST-912: allows temp invoices to be deleted, and will remove paperclip (attachments).

If the temp invoice has not been sent (ie the Document is not attached to anything else), then will also delete the Document itself.

Also, moves the functionality for creating email and postal comms from Document mixins to equivalent mixins on Invoice.  (These just delegate to the Document_xxx mixins under the covers).  The main benefit of this is finer-grained control as to the handling of attachments: when the Document is initially created, it is only attached to the Invoice.  But, when the Invoice is sent (by either Invoice_email or Invoice_print), then additional attachments for the invoice's buyer and seller are created (in addition to that of the comm).

If the temp invoice is deleted after it has been sent, the attachment to the buyer, seller and comm remain (even though the temp invoice's paperclip is deleted along with the temp invoice itself).

EST-912: adds integration test for creating document directly from Invoice (rather than InvoiceSummary)

EST-912: consolidates document management tests for Invoice into a single integtest class, simplifies and extends the test for Invoice#remove to check that the Document is also deleted.

In addition, updates all pom with standard assertj dependencies.

EST-889: adds test to verify that if an invoice is removed with a document that has been sent, then the document is NOT removed and the paperclips (to invoice's buyer, seller and also the new comm) are all preserved.

In addition, moved the runBackgroundCommands helper method up to EstatioIntegrationTest so can be reused across tests.

EST-888: adds test that can only delete documents that have not been sent.

EST-917: prevents being able to prepare an invoice note for an invoice that is still changeable (eg hasn't been invoiced)

EST-917: adds integ test to confirm that cannot create an invoice note directly for an invoice that has not yet been invoiced

EST-910: hides Paperclip_changeRole mixin, so can only be chaned programmatically.

This is required because the role names are well-defined and fixed

EST-928: converts Invoice's actions to be mixins

EST-908: changes for splitting of document module's Binder interface into RendererModelFactory and AttachmentAdvisor

EST-908: reworks against 1.13.10 of document module (RendererModelFactory and AttachmentAdvisor interfaces)

EST-908: when attach receipt to an invoice that already has an invoiceDoc associated with it, then copy over.  Also cascade delete of Paperclip_Document if a document is deleted.  In addition...

In addition:
- PaperclipRoleNames to catalog all names in single spot.
-moved DocAndCommService and renamed to InvoiceDocAndCommService

EST-908: renames DocAndCommForInvoiceNote  to DocAndCommForInvoiceDoc instead (ubiquitous language).

EST-908: adds mixins for prepare/sendByEmail/sendByPost for the DnC view models on Invoice itself (same UI as the InvoiceSummary).  Also..

... also, completed the renaming of InvoiceNote -> InvoiceDoc.

EST-926: mixins for Document (from the Communication module) to show the comms that have been sent (or attachments that would be sent).

Also updating the -ITA email cover note, as was corrupted somehow.

EST-926: updates to 1.13.11 of document module