EST-1387: allows user to specify person to assign to for all the relevant transitions in the incoming invoice approval cycle
specifically:
- complete
- approve

None of the other transitions (this includes "approve by director") is followed by a transition that has an associated task, so there's no point in prompting.

For these triggers, now also shows the role that the person belongs to
(admittedly, as an unused string.  Use jQuery (in ECP project) to make uneditable).

In addition:
- fixed guards/transitions for trigger actions
- removed the _nextApprovalTaskPersonAssignedTo mixin + subtypes
- removed the mixins on Task for _nextCategorisationTask (no point really, and hadn't been done for other 3 types)
- removed the _reasonApprovalTaskBlocked mixin - instead we disable the trigger actions

Also:
- restricts 2 prototype actions ("bad command targets" and "open swagger") so not visible in production mode