EST-1541: flushes transaction after deleting paperclips.
There are two subscribers.  The one that zaps tasks and transitions flushes the cache, but the one for paperclips did not.  The order in which subscribers are called is not deterministic, so if the paperclip subscriber was called before the tasks and transition one, then all would be good, but the other way around and the error would occur.

Fix is to flush the transaction in the paperclips subscriber also.