EST-1554: when calculating invoiced total on project, invoice items that are not linked to a budget also contribute to the total; also improves display of amounts
For discussion:
I am not sure what the users expect. The current implementation works bottom up from order items. It inspects the invoiced amount on an order item and adds it to the total invoiced on an project item.
So invoice items that have a project but are not linked to an order item are ‘ignored’. Another implementation would be to look at the combination of project-charge on invoice item.