EST-1654: completedBy on transitions only populated for explicit actions
two subtleties here:
a) an invoice might have been rejected and not yet been fixed (ie completed).  In this case the approval transition must be ignored.
b) an invoice might be in approved but was rejected and then fixed.  In this case the most recent "approval" state transition will have a null completed by, so need to skip and look for the original approval.

related:
- for consistency/ease of processing, store just the username "fbloggs@ecp.loc" rather than the name "Fred Bloggs (fbloggs@ecp.loc)"
  - correspond flyway migration script provided in ecp repo