EST-1657: "tasks for me" now also shows tasks assigned to no-one.
As a refinement on the previous algorithm, we no longer assign a task to an individual Person if there is >1 person that has the role (eg treasurer role).  If there is exactly one person, though, then we do assign the task.

On the home page, "tasks for me" (LHS) also lists tasks assigned to no-one, but for which the current user has the roles to perform