EST-1854: REMOVES SYNC - the capability to sync Party, BankAccount or BankMandate
because the syncing of element 6 info happens as a side-effect of collect/invoice on the InvoiceDto, and there sync'ing of Party/BankAccount/BankMandate in themselves is not (no longer) safe as we don't check if they are only a tenant or only a supplier.

We can also get rid of the DtoFactory for PartyBankAccountsAndBankMandates.