EST-1862: adds capability to disable autosync of valid CodaDocHeads
Main change is that DerivedObjectUpdater, rather than looking at whether the CodaDocHead is valid, instead asks if its handling is set ot sync.  This means moving the responsibility for setting handling earlier, into the CodaDocHead's own validateOnly method.  For estatio-ecp, the corresponding method in CodaDocService (updateSyncAndHandling) is inlined because its only responsibility is to set the invoice on the CodaDocHead.