ECP-1095: no longer caches CodaDocHead#analysisLineDataByLineNumber ... since NOT IDEMPOTENT
this is where our issue was coming from, I believe.

This method CodaDocHead#getAnalysisLineDataByLineNumber(), is only called by the Memento code.

Normally the CodaDocHead entity will be discarded by the runtime at the end of the session, and so there's no real problem with the Map being transient - it's just a little performance optimization.

However, in our race condition we have:
- CodaDocService#handle(..) called twice in quick succession via the REST API
- each time its called they - in the old design - delegated the syncIfValid stuff off to the background scheduler
- the background scheduler runs once every 10 seconds, and so finds the two commands to sync for the same target CodaDocHead and dispatches them in quick succession, but - this is the important bit - it does each in a separate transaction but in the same session

for the first sync, the CodaDocHead creates the map for the first time, and the LineData will not have an associated invoice item - because first time there is not yet any invoice.  The sync'ing then continues and ends up creating an Invoice + InvoiceItem
for the second sync, this will reuse the same CodaDocHead entity, because in same DN session; trouble is this will still have the map populated, and this map is incorrect because the LineData for the first will not reference the InvoiceItem just created.

What happens then is that the sync ends up creating another item.

Doh !

Since this method is only ever called by Memento, and Memento itself is an immutable snapshot for a point in time, our fix is simply to get rid of the caching and recompute each time.

Subsequent commits will simplify this code.